Give the bridge server factory explicit file ownership
Move the existing public factory unchanged into its matching infrastructure file so the remaining server source can focus on listener and session lifecycle responsibilities.
