+++ Packages/src/Editor/Infrastructure/UnityCliLoopBridgeServerInstanceFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 620d1d36ae77447990654fa185c55822
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
